Add __format__ magic method
__format__ method is implemented to support native string formatting.
Also __call__ is implemented to get name to determine suffix.
It works like this: first we get the name to be formatted with __call__
__call__ returns the instance itself but the state of instance has changed.
After that __format__ is called by str.format with name is known. The type of
suffix is determined from format_spec. The last 3 characters are considered for
determination of suffix.